v0.20.3
- Artillery Enemy
 - Can launch projectiles at 5 positions.
 - 5 positions are not linked due to instantiated prefabs. NEED to link it.

 - Time to add draggable consumable items.

+++ Cattler/Assets/1.MainProj/Script/EnemyUnit.cs
@@ -19,7 +19,7 @@ public class EnemyUnit : MonoBehaviour
     [Header("Stats")]
     [SerializeField] public int currentHealth;
     [SerializeField] public float attackSpeed;
-    [SerializeField] public float attackDamage;
+    [SerializeField] public int attackDamage;
     [SerializeField] public float moveSpeed;
     [SerializeField] public float attackRange;
